+++ dubbo-samples-api/src/main/java/org/apache/dubbo/samples/api/GreetingsService.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.dubbo.samples.api;
+package org.apache.dubbo.samples.api;
 
 import com.alibaba.dubbo.config.async.DubboAsync;
