+++ dubbo-samples-api/src/main/java/org/apache/dubbo/samples/server/GreetingsServiceImpl.java
@@ -15,22 +15,13 @@
  * limitations under the License.
  */
 
-package com.alibaba.dubbo.samples.client;
+package org.apache.dubbo.samples.server;
 
-import com.alibaba.dubbo.samples.api.GreetingsService;
+import org.apache.dubbo.samples.api.GreetingsService;
 
-import org.apache.dubbo.config.ApplicationConfig;
-import org.apache.dubbo.config.ReferenceConfig;
-import org.apache.dubbo.config.RegistryConfig;
-
-public class Application {
-    public static void main(String[] args) {
-        ReferenceConfig<GreetingsService> reference = new ReferenceConfig<>();
-        reference.setApplication(new ApplicationConfig("first-dubbo-client"));
-        reference.setRegistry(new RegistryConfig("multicast://224.5.6.7:1234"));
-        reference.setInterface(GreetingsService.class);
-        GreetingsService greetingsService = reference.get();
-        String message = greetingsService.sayHi("dubbo");
-        System.out.println(message);
+public class GreetingsServiceImpl implements GreetingsService {
+    @Override
+    public String sayHi(String name) {
+        return "hi, " + name;
     }
 }
